move 'Add Service' button on the top of EditTrip page
